feat(plugin-stripe): adopt v4 design tokens and restyle LinkToDoc field for v4 ui
