+++ packages/plugin-stripe/src/ui/LinkToDoc/index.css
@@ -1,0 +1,26 @@
+@layer payload-default {
+  .link-to-doc {
+    margin-bottom: var(--spacing-field);
+  }
+
+  .link-to-doc__header {
+    display: flex;
+    align-items: center;
+    width: fit-content;
+    gap: var(--spacer-1);
+  }
+
+  .link-to-doc__url {
+    overflow: hidden;
+    font-weight: 600;
+    text-overflow: ellipsis;
+  }
+
+  .link-to-doc__url a {
+    color: var(--color-text-secondary);
+
+    &:hover {
+      color: var(--color-text);
+    }
+  }
+}
